refactor: simplify validation code for better maintainability
- Consolidate all validation logic into validate.go
- Remove duplicate shell_safe.go file
- Merge validation functions into single validateShellSafe function
- Simplify linter by removing ast-grep dependency
- Keep all security checks while reducing code complexity

All tests pass and validation functionality remains intact.
