+++ tools/lint-unsafe-templates.sh
@@ -12,66 +12,34 @@ NC='\033[0m' # No Color
 
 echo "=== Checking for unsafe template usage ==="
 
-# Create temporary file for ast-grep rules
-RULES_FILE=$(mktemp)
-trap "rm -f $RULES_FILE" EXIT
+# Simple linter that checks for basic patterns
+echo -e "${YELLOW}Checking for direct template embedding patterns...${NC}"
 
-# Rule 1: Direct use of deref without validation in shell templates
-cat > "$RULES_FILE" <<'EOF'
-rules:
-  - id: direct-deref-in-template
-    pattern: |
-      {{ deref $ARG }}
-    message: "Direct use of deref in template without explicit validation"
-    severity: warning
-    note: |
-      The deref function is being used directly in a template.
-      Ensure that the dereferenced value has been validated for shell safety.
-      Consider using ValidateAllFields() before template rendering.
-EOF
+# Look for patterns that might embed user input directly without validation
+PATTERNS=(
+    "{{ *deref[^}]*}}"
+    "BINARY_NAME_[0-9]="
+    "BINARY_PATH_[0-9]="
+    "OS=.*deref"
+    "ARCH=.*deref"
+    "NAME=.*deref"
+    "REPO=.*deref"
+)
 
-# Check for ast-grep
-if ! command -v ast-grep &> /dev/null; then
-    echo -e "${YELLOW}Warning: ast-grep not found. Install with: brew install ast-grep${NC}"
-    echo "Falling back to grep-based checks..."
-
-    # Fallback: Use grep to find potential issues
-    echo -e "\n${YELLOW}Checking for direct template embedding patterns...${NC}"
-
-    # Look for patterns that might embed user input directly
-    PATTERNS=(
-        "{{ *deref[^}]*}}"
-        "BINARY_NAME_[0-9]="
-        "BINARY_PATH_[0-9]="
-        "OS=.*deref"
-        "ARCH=.*deref"
-        "NAME=.*deref"
-        "REPO=.*deref"
-    )
-
-    FOUND_ISSUES=false
-    for pattern in "${PATTERNS[@]}"; do
-        echo -e "\nChecking for pattern: $pattern"
-        if grep -r -n -E "$pattern" internal/shell/template.tmpl.sh 2>/dev/null; then
-            FOUND_ISSUES=true
-        fi
-    done
-
-    if [ "$FOUND_ISSUES" = true ]; then
-        echo -e "\n${YELLOW}⚠ Found potential unsafe template patterns${NC}"
-        echo "These patterns embed user input directly into shell scripts."
-        echo "Ensure ValidateAllFields() is called before template rendering."
-    else
-        echo -e "\n${GREEN}✓ No unsafe template patterns found${NC}"
+FOUND_ISSUES=false
+for pattern in "${PATTERNS[@]}"; do
+    echo -e "\nChecking for pattern: $pattern"
+    if grep -r -n -E "$pattern" internal/shell/template.tmpl.sh 2>/dev/null; then
+        FOUND_ISSUES=true
     fi
+done
+
+if [ "$FOUND_ISSUES" = true ]; then
+    echo -e "\n${YELLOW}⚠ Found potential unsafe template patterns${NC}"
+    echo "These patterns embed user input directly into shell scripts."
+    echo "Ensure ValidateAllFields() is called before template rendering."
 else
-    # Use ast-grep for more sophisticated checking
-    echo "Running ast-grep checks..."
-
-    # Check Go template files
-    if ast-grep --rule-file "$RULES_FILE" internal/shell/template.tmpl.sh 2>/dev/null; then
-        echo -e "${YELLOW}⚠ Found potential unsafe template usage${NC}"
-    fi
+    echo -e "\n${GREEN}✓ No unsafe template patterns found${NC}"
 fi
 
 # Additional checks using standard tools
@@ -100,5 +68,4 @@ fi
 echo -e "\n=== Lint Summary ==="
 echo "1. Always call ValidateAllFields() before generating scripts"
 echo "2. Never embed user input directly without validation"
-echo "3. Use shellQuote for dynamic values when necessary"
-echo "4. Review any new template functions for security"
+echo "3. Review any new template functions for security"
